feat: introduce cobra command framework
