+++ go.mod
@@ -2,10 +2,15 @@ module github.com/MatthiasHarzer/patreon-crawler
 
 go 1.26.2
 
-require github.com/stretchr/testify v1.11.1
+require (
+	github.com/spf13/cobra v1.10.2
+	github.com/stretchr/testify v1.11.1
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.9 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
